funkcja kwadratowa v4

+++ FunkcjaKwadratowaTest/FunkcjaKwadratowaTest/Obliczanie_miejsc_zerowych.cs
@@ -9,6 +9,16 @@ namespace FunkcjaKwadratowaTest
 {
     public class Obliczanie_miejsc_zerowych
     {
+        // Tego się nie da zwobić w obecnej formie funcji Wylicz - trzeba zmienić na IEnumerable
+        //[Theory,
+        //    InlineData(1, -4, 3, new double[] { 1, 3 }),
+        //    InlineData(9, -12, 4, new double[] { 2.0 / 3 }),
+        //    InlineData(-6, 3, -1, new double[] { })]
+        //public void Obliczanie_miejsc_zerowych_wszystkie_przypadki(double a, double b, double c, double[] oczekiwane)
+        //{
+        //    var wynik = Oblicz(a, b, c);
+        //    Assert.Equal(oczekiwane, wynik);
+        //}
         [Fact]
         public void Potrafimy_znalezc_dwa_miejsca_zerowe()
         {
@@ -42,7 +52,7 @@ namespace FunkcjaKwadratowaTest
             Assert.Equal(double.NaN, wynik.MiejscePierwsze);
             Assert.Equal(double.NaN, wynik.MiejsceDrugie);
         }
-        private Wynik Oblicz(int a, int b, int c)
+        private Wynik Oblicz(double a, double b, double c)
         {
             double delta = b * b - 4 * a * c;
             double x1;
